use time slots controller in main route
